Draugr an
Thukker Hardsuit had less proc than even the base SHI HEV. Buffed it to a very reasonable 15.

Al'Seik hardsuit was incapable of entering melee combat previously as it's health was too low and its proc was too low as well. Meaning one of the main gimmicks of TAP couldn't be used without copious amounts of stimulants just to get close.